pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codetraceai"
version = "0.1.0"
description = "AI 编程辅助与知识沉淀工具"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "CodeTraceAI Team"}
]
keywords = ["ai", "coding", "assistant", "knowledge", "claude"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "anthropic>=0.18.0",
    "openai>=1.0.0",
    "gitpython>=3.1.40",
    "pysvn>=1.9.0",
    "PySide6>=6.6.0",
    "PyQt-Fluent-Widgets>=1.5.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "keyring>=24.0.0",
    "cryptography>=41.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
codetrace = "codetrace.cli:cli"

[project.urls]
Homepage = "https://github.com/yourusername/codetraceai"
Repository = "https://github.com/yourusername/codetraceai"
Issues = "https://github.com/yourusername/codetraceai/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "codetrace*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
